make currency converter

+++ app/src/main/java/ru/akumakeito/currention/ui/screens/CurrencyConverterScreen.kt
@@ -53,7 +53,7 @@ fun CurrencyConverterScreen(
                 onAmountTextChanged = { amount ->
                     coroutineScope.launch {
                         delay(3000)
-                        convertCurrencyViewModel.changeAmount(amount.toDouble())
+                        convertCurrencyViewModel.changeAmount(amount)
                     }
